webEEGPortal/requirements.txt
absl-py==0.7.1
alabaster==0.7.12
amqp==2.4.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
astropy==3.1.2
asyncio==3.4.3
asyncssh==1.16.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
billiard==3.6.0.0
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
Bottleneck==1.2.1
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
colorama==0.4.1
contextlib2==0.5.5
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
dask==1.1.4
decorator==4.4.0
defusedxml==0.5.0
dicom==0.9.9.post1
distributed==1.26.0
Django==2.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-headers==2.5.2
django-require==1.0.11
django-storages==1.7.1
django-timezone-field==3.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==1.0.2
flower==0.9.3
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.20.0
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.0.0
invoke==1.2.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
isort==4.3.16
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
jeepney==0.4
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==18.0.0
kiwisolver==1.0.1
kombu==4.5.0
lazy-object-proxy==1.3.1
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.2
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
med2image==1.1.2
mistune==0.8.4
mock==2.0.0
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
mysqlclient==1.4.2.post1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
neurdflib==5.0.0.post1
nibabel==2.4.0
nipype==1.1.9
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
olefile==0.46
opencv-python==4.1.0.25
openpyxl==2.6.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.3.4
partd==0.3.10
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pbr==5.1.3
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.9.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
prov==1.5.3
psutil==5.6.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pydicom==1.2.2
pydot==1.4.1
pydotplus==2.0.2
pyflakes==2.1.1
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-crontab==2.3.6
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
rdflib==4.2.2
redis==3.2.1
requests==2.21.0
rope==0.12.0
ruamel-yaml==0.15.46
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
sqlparse==0.3.0
statsmodels==0.9.0
sympy==1.3
tables==3.5.1
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.4
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
traits==5.1.1
typed-ast==1.3.1
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
zict==0.1.4
zipp==0.3.3
